Add files via upload
Selection sort algorithm works by dividing the original array in two subarrays: Sorted subarray and unsorted subarray, initially the sorted subarray is empty. This algorithm works by repeatedly finding the minimum element from the unsorted subarray and replacing it with the first element of the array, thus making that part of the array as sorted subarray. This happens repeatedly until the whole array is sorted